+++ Structural/Adapter/Adapter.csproj
