+++ Structural/Adapter/AdapterMain.cs
