+++ Structural/Adapter/ComponentPdf.cs
