+++ Structural/Adapter/Document.cs
