+++ Structural/Adapter/DocumentPdf.cs
