+++ Structural/Adapter/Documenthtml.cs
