+++ Structural/Adapter/Properties/AssemblyInfo.cs
