+++ Structural/Bridge/Properties/AssemblyInfo.cs
