+++ Structural/Composite/Company.cs
