+++ Structural/Composite/CompanyWithoutSubsidiary.cs
