+++ Structural/Composite/Composite.csproj
