+++ Structural/Composite/CompositeMain.cs
