+++ Structural/Composite/ParentCompany.cs
