+++ Structural/Composite/Properties/AssemblyInfo.cs
