+++ Structural/Decorator/DecoratorForBrand.cs
