+++ Structural/Decorator/DecoratorForModel.cs
