+++ Structural/Decorator/DecoratorMain.cs
