+++ Structural/Decorator/ICarDisplayComponent.cs
