+++ Structural/Decorator/Properties/AssemblyInfo.cs
