+++ Structural/Facade/CatalogComponent.cs
