+++ Structural/Facade/DocumentManagementComponent.cs
