+++ Structural/Facade/ICatalog.cs
