+++ Structural/Facade/IDocumentManagement.cs
