+++ Structural/Facade/Properties/AssemblyInfo.cs
