+++ Structural/FlighWeight/OptionsFactory.cs
