+++ Structural/FlighWeight/Properties/AssemblyInfo.cs
