+++ Structural/Proxy/AnimationProxy.cs
