+++ Structural/Proxy/IAnimationMovie.cs
