+++ Structural/Proxy/Properties/AssemblyInfo.cs
